+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ----------
 
+- add --install option to test command
+- allow backslash line continuations in recipe files
+
 0.4.51 [build edbd5b]
 ---------------------